--- src/app/(frontend)/(public)/home/layout.tsx
@@ -1,7 +1,0 @@
-export default function HomeLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <section className="bg-gray-100">{children}</section>;
-}
