+++ src/app/components/Advantages/Advantages.tsx
