+++ src/app/components/CardAdv/CardAdv.tsx
