+++ src/app/components/NavBar/NavBar.tsx
@@ -57,7 +57,7 @@ function NavBar() {
         <div 
           className={`${
             menuOpen ? 'flex' : 'hidden'
-          } ${isScrolled ? 'md:mt-5 mt-0 md:rounded-xl rounded-b-3xl shadow-2xl transition-all duration-500' : 'mt-0 rounded-b-2xl transition-all duration-500'} md:flex flex-col items-end absolute pr-7 pb-7 pt-3 gap-8 w-screen text-base md:text-sm lg:text-base font-semibold md:static md:gap-0 md:p-0 md:flex-row md:justify- md:w-2/4 md:h-10 bg-white md:font-bold`}
+          } ${isScrolled ? 'md:mt-5 mt-0 md:rounded-xl rounded-b-3xl shadow-2xl transition-all duration-500' : 'mt-0 rounded-b-2xl transition-all duration-500'} md:flex flex-col items-end absolute pr-7 pb-7 pt-3 gap-8 w-screen text-base md:text-sm lg:text-base font-semibold md:static md:gap-0 md:p-0 md:flex-row md:justify- md:w-2/4 md:h-10 bg-white text-black md:font-bold`}
         >
           <div onClick={handleHomeClick} className='md:h-full md:w-1/5 md:flex justify-center items-center cursor-pointer hover:text-green_greenplus hover:rounded-l-xl transition-colors duration-300'>
             <p>Home</p>
