+++ src/app/components/ProjectsCarousel/ProjectsCarousel.tsx
