+++ src/app/components/WaveEffect/WaveEffect.tsx
