+++ src/app/page.tsx
@@ -1,13 +1,25 @@
+import Contact from "./components/Contact/Contact"
+import Footer from "./components/Footer/Footer"
+import Headboard from "./components/Headboard/Headboard"
+import OurProducts from "./components/OurProjects/OurProjects"
+import Team from "./components/Team/Team"
+import OurServices from "./components/OurServices/OurServices"
+import Advantages from "./components/Advantages/Advantages"
 import Link from "next/link";
 import "./Home.css";
 import "./globals.css";
 
-export default async function RootLayout() {
-	return ( 
-		<>
-			Inicio
-			<br />
-			<Link href="/private/dashboard">Dashboard</Link>
-		</> 
-	);
+function Home() {
+  return (
+    <div>
+      <Headboard />
+      <OurServices />
+      <Advantages />
+      <Team />
+      <OurProducts />
+      <Contact />
+      <Footer />
+    </div>
+  )
 }
+export default Home
